+++ src/collective/contact/plonegroup/tests/test_settings.py
@@ -2,14 +2,17 @@
 """Setup/installation tests for this package."""
 
 from collective.contact.plonegroup.browser import settings
-from collective.contact.plonegroup.config import FUNCTIONS_REGISTRY
-from collective.contact.plonegroup.config import ORGANIZATIONS_REGISTRY
+from collective.contact.plonegroup.config import DEFAULT_DIRECTORY_ID
 from collective.contact.plonegroup.config import PLONEGROUP_ORG
+from collective.contact.plonegroup.config import get_registry_functions
+from collective.contact.plonegroup.config import get_registry_organizations
+from collective.contact.plonegroup.config import set_registry_functions
+from collective.contact.plonegroup.config import set_registry_organizations
 from collective.contact.plonegroup.testing import IntegrationTestCase
+from collective.contact.plonegroup.utils import get_own_organization
 from collective.contact.plonegroup.utils import get_plone_group_id
 from plone import api
 from plone.app.testing import TEST_USER_ID
-from plone.registry.interfaces import IRegistry
 from z3c.form import validator
 from zExceptions import Redirect
 from zope import event
@@ -18,7 +21,6 @@ from zope.i18n import translate
 from zope.interface import Invalid
 from zope.lifecycleevent import ObjectModifiedEvent
 from zope.schema.interfaces import IVocabularyFactory
-from copy import deepcopy
 
 
 class TestSettings(IntegrationTestCase):
@@ -28,9 +30,9 @@ class TestSettings(IntegrationTestCase):
         """Custom shared utility setup for tests."""
         self.portal = self.layer['portal']
         # Organizations creation
-        self.portal.invokeFactory('directory', 'contacts')
-        self.portal['contacts'].invokeFactory('organization', PLONEGROUP_ORG, title='My organization')
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        self.portal.invokeFactory('directory', DEFAULT_DIRECTORY_ID)
+        self.portal[DEFAULT_DIRECTORY_ID].invokeFactory('organization', PLONEGROUP_ORG, title='My organization')
+        own_orga = get_own_organization()
         own_orga.invokeFactory('organization', 'department1', title='Department 1')
         own_orga.invokeFactory('organization', 'department2', title='Department 2')
         own_orga['department1'].invokeFactory('organization', 'service1', title='Service 1')
@@ -38,18 +40,15 @@ class TestSettings(IntegrationTestCase):
         inactive_department = own_orga['inactive_department']
         api.content.transition(obj=inactive_department, transition='deactivate')
 
-        self.registry = getUtility(IRegistry)
-        self.registry['collective.contact.plonegroup.browser.settings.IContactPlonegroupConfig.'
-                      'organizations'] = [own_orga['department1'].UID(),
-                                          own_orga['department1']['service1'].UID(),
-                                          own_orga['department2'].UID()]
-        self.registry['collective.contact.plonegroup.browser.settings.IContactPlonegroupConfig.'
-                      'functions'] = [{'fct_title': u'Director',
-                                       'fct_id': u'director',
-                                       'fct_orgs': []},
-                                      {'fct_title': u'Worker',
-                                       'fct_id': u'worker',
-                                       'fct_orgs': []}]
+        set_registry_organizations([own_orga['department1'].UID(),
+                                   own_orga['department1']['service1'].UID(),
+                                   own_orga['department2'].UID()])
+        set_registry_functions([{'fct_title': u'Director',
+                                 'fct_id': u'director',
+                                 'fct_orgs': []},
+                                {'fct_title': u'Worker',
+                                 'fct_id': u'worker',
+                                 'fct_orgs': []}])
 
     def test_OwnOrganizationServicesVocabulary(self):
         """ Test vocabulary """
@@ -59,16 +58,16 @@ class TestSettings(IntegrationTestCase):
         self.assertSetEqual(set(voc_list), set(['Department 1 - Service 1', 'Department 1', 'Department 2']))
         self.assertNotIn('Inactive department', voc_list)
         # When multiple own organizations
-        self.portal['contacts'].invokeFactory('organization', 'temporary', title='Temporary')
-        self.portal['contacts']['temporary'].invokeFactory('organization', PLONEGROUP_ORG,
-                                                           title='Duplicated organization')
+        self.portal[DEFAULT_DIRECTORY_ID].invokeFactory('organization', 'temporary', title='Temporary')
+        self.portal[DEFAULT_DIRECTORY_ID]['temporary'].invokeFactory(
+            'organization', PLONEGROUP_ORG, title='Duplicated organization')
         services = getUtility(IVocabularyFactory, name=u'collective.contact.plonegroup.organization_services')
         voc_dic = services(self).by_token
         voc_list = [voc_dic[key].title for key in voc_dic.keys()]
         self.assertEquals(voc_list, [u"You must have only one organization with id '${pgo}' !"])
-        self.portal['contacts'].manage_delObjects(ids=['temporary'])
+        self.portal[DEFAULT_DIRECTORY_ID].manage_delObjects(ids=['temporary'])
         # When own organization not found
-        self.portal['contacts'].manage_delObjects(ids=[PLONEGROUP_ORG])
+        self.portal[DEFAULT_DIRECTORY_ID].manage_delObjects(ids=[PLONEGROUP_ORG])
         services = getUtility(IVocabularyFactory, name=u'collective.contact.plonegroup.organization_services')
         voc_dic = services(self).by_token
         voc_list = [voc_dic[key].title for key in voc_dic.keys()]
@@ -77,7 +76,7 @@ class TestSettings(IntegrationTestCase):
     def test_detectContactPlonegroupChange(self):
         """Test if group creation works correctly"""
         group_ids = [group.id for group in api.group.get_groups()]
-        organizations = self.registry[ORGANIZATIONS_REGISTRY]
+        organizations = get_registry_organizations()
         for uid in organizations:
             self.assertIn('%s_director' % uid, group_ids)
             self.assertIn('%s_worker' % uid, group_ids)
@@ -86,29 +85,28 @@ class TestSettings(IntegrationTestCase):
         d1s1_d_group = api.group.get(groupname='%s_director' % organizations[1])
         self.assertEquals(d1s1_d_group.getProperty('title'), 'Department 1 - Service 1 (Director)')
         # Changing function title
-        self.registry[FUNCTIONS_REGISTRY] = [{'fct_title': u'Directors', 'fct_id': u'director', 'fct_orgs': []},
-                                             {'fct_title': u'Worker', 'fct_id': u'worker', 'fct_orgs': []}]
+        set_registry_functions([{'fct_title': u'Directors', 'fct_id': u'director', 'fct_orgs': []},
+                                {'fct_title': u'Worker', 'fct_id': u'worker', 'fct_orgs': []}])
         d1_d_group = api.group.get(groupname='%s_director' % organizations[0])
         self.assertEquals(d1_d_group.getProperty('title'), 'Department 1 (Directors)')
         d1s1_d_group = api.group.get(groupname='%s_director' % organizations[1])
         self.assertEquals(d1s1_d_group.getProperty('title'), 'Department 1 - Service 1 (Directors)')
         # Adding new organization
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         own_orga['department2'].invokeFactory('organization', 'service2', title='Service 2')
         # append() method on the registry doesn't trigger the event. += too
-        newValue = self.registry[ORGANIZATIONS_REGISTRY] + [own_orga['department2']['service2'].UID()]
-        self.registry[ORGANIZATIONS_REGISTRY] = newValue
+        newValue = get_registry_organizations() + [own_orga['department2']['service2'].UID()]
+        set_registry_organizations(newValue)
         group_ids = [group.id for group in api.group.get_groups()]
-        last_uid = self.registry[ORGANIZATIONS_REGISTRY][-1]
+        last_uid = get_registry_organizations()[-1]
         self.assertIn('%s_director' % last_uid, group_ids)
         self.assertIn('%s_worker' % last_uid, group_ids)
         # Adding new function
-        newValue = self.registry[FUNCTIONS_REGISTRY] + [
-            {'fct_title': u'Chief', 'fct_id': u'chief', 'fct_orgs': []}]
-        self.registry[FUNCTIONS_REGISTRY] = newValue
+        newValue = get_registry_functions() + [{'fct_title': u'Chief', 'fct_id': u'chief', 'fct_orgs': []}]
+        set_registry_functions(newValue)
         group_ids = [group.id for group in api.group.get_groups() if '_' in group.id]
         self.assertEquals(len(group_ids), 12)
-        for uid in self.registry[ORGANIZATIONS_REGISTRY]:
+        for uid in get_registry_organizations():
             self.assertIn('%s_director' % uid, group_ids)
             self.assertIn('%s_chief' % uid, group_ids)
             self.assertIn('%s_worker' % uid, group_ids)
@@ -116,21 +114,21 @@ class TestSettings(IntegrationTestCase):
     def test_detectContactPlonegroupChangeRemoveFunction(self):
         """When a function is removed, every linked Plone groups are deleted as well.
            This is protected by validateSettings that checks first that every Plone groups are empty."""
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         dep1 = own_orga['department1']
         plone_group_id = get_plone_group_id(dep1.UID(), 'director')
         self.assertTrue(api.group.get(plone_group_id))
-        functions = list(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         # remove 'director'
         functions.pop(0)
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         # the linked Plone groups are deleted
         self.assertFalse(api.group.get(plone_group_id))
 
     def test_detectContactPlonegroupChangeSelectOrgs(self):
         """When selecting 'fct_orgs' on a function, Plone groups are create/deleted depending
            on the fact that 'fct_orgs' is empty or contains some organization uids."""
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         dep1 = own_orga['department1']
         dep1_uid = dep1.UID()
         dep2 = own_orga['department2']
@@ -140,42 +138,42 @@ class TestSettings(IntegrationTestCase):
         self.assertTrue(api.group.get(dep1_plone_group_id))
         self.assertTrue(api.group.get(dep2_plone_group_id))
         # select dep2_uid for 'director'
-        functions = deepcopy(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         functions[0]['fct_orgs'] = [dep2_uid]
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         # dep1 director Plone group is deleted
         self.assertFalse(api.group.get(dep1_plone_group_id))
         self.assertTrue(api.group.get(dep2_plone_group_id))
         # select dep1_uid for 'director'
-        functions = deepcopy(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         functions[0]['fct_orgs'] = [dep1_uid]
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         self.assertTrue(api.group.get(dep1_plone_group_id))
         self.assertFalse(api.group.get(dep2_plone_group_id))
         # select nothing for 'director', every groups are created
-        functions = deepcopy(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         functions[0]['fct_orgs'] = []
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         self.assertTrue(api.group.get(dep1_plone_group_id))
         self.assertTrue(api.group.get(dep2_plone_group_id))
         # select both dep1 and dep2, every groups are created
-        functions = deepcopy(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         functions[0]['fct_orgs'] = [dep1_uid, dep2_uid]
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         self.assertTrue(api.group.get(dep1_plone_group_id))
         self.assertTrue(api.group.get(dep2_plone_group_id))
 
     def test_validateSettingsRemoveFunction(self):
         """A function may only be removed if every linked Plone groups are empty."""
         # add a user to group department1 director
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         dep1 = own_orga['department1']
         plone_group_id = get_plone_group_id(dep1.UID(), 'director')
         api.group.add_user(groupname=plone_group_id, username=TEST_USER_ID)
         invariants = validator.InvariantsValidator(
             None, None, None, settings.IContactPlonegroupConfig, None)
-        orgs = list(self.registry[ORGANIZATIONS_REGISTRY])
-        functions = list(self.registry[FUNCTIONS_REGISTRY])
+        orgs = get_registry_organizations()
+        functions = get_registry_functions()
         data = {'organizations': orgs, 'functions': functions}
         # for now it validates correctly
         self.assertFalse(invariants.validate(data))
@@ -198,15 +196,15 @@ class TestSettings(IntegrationTestCase):
            is only possible if groups that will be deleted (Plone groups of organizations not selected
            as 'fct_orgs') are empty."""
         # add a user to group department1 director
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         dep1 = own_orga['department1']
         dep2 = own_orga['department2']
         plone_group_id = get_plone_group_id(dep1.UID(), 'director')
         api.group.add_user(groupname=plone_group_id, username=TEST_USER_ID)
         invariants = validator.InvariantsValidator(
             None, None, None, settings.IContactPlonegroupConfig, None)
-        orgs = deepcopy(self.registry[ORGANIZATIONS_REGISTRY])
-        functions = deepcopy(self.registry[FUNCTIONS_REGISTRY])
+        orgs = get_registry_organizations()
+        functions = get_registry_functions()
         data = {'organizations': orgs, 'functions': functions}
         # set dep2 as 'fct_orgs' of 'director' function
         director = functions[0]
@@ -225,8 +223,8 @@ class TestSettings(IntegrationTestCase):
 
     def test_adaptPloneGroupDefinition(self):
         """ Test event when an organization is changed """
-        organizations = self.registry[ORGANIZATIONS_REGISTRY]
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        organizations = get_registry_organizations()
+        own_orga = get_own_organization()
         # an organization is modified
         own_orga['department1'].title = 'Department 1 changed'
         event.notify(ObjectModifiedEvent(own_orga['department1']))
@@ -252,15 +250,15 @@ class TestSettings(IntegrationTestCase):
     def test_onlyRelevantPloneGroupsCreatedWhenFunctionRestrictedToSelectedOrgs(self):
         """Test using 'fct_orgs' when defining functions."""
         # create a new suffix and restrict it to department1
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        own_orga = get_own_organization()
         dep1 = own_orga['department1']
         dep1_uid = dep1.UID()
         dep2 = own_orga['department2']
         dep2_uid = dep2.UID()
-        functions = list(self.registry[FUNCTIONS_REGISTRY])
+        functions = get_registry_functions()
         new_function = {'fct_id': u'new', 'fct_title': u'New', 'fct_orgs': [dep1_uid]}
         functions.append(new_function)
-        api.portal.set_registry_record(FUNCTIONS_REGISTRY, functions)
+        set_registry_functions(functions)
         # 'new' suffixed Plone group was created only for dep1
         self.assertTrue(api.group.get(get_plone_group_id(dep1_uid, u'new')))
         self.assertFalse(api.group.get(get_plone_group_id(dep2_uid, u'new')))
@@ -291,12 +289,11 @@ class TestSettings(IntegrationTestCase):
         self.assertEqual(len(vocab_all), 3)
         self.assertListEqual([v.title for v in vocab_all],
                              ['Department 1', 'Department 1 - Service 1', 'Department 2'])
-        registry = getUtility(IRegistry)
-        registry[ORGANIZATIONS_REGISTRY] = [vocab_all_values[2], vocab_all_values[0]]
+        set_registry_organizations([vocab_all_values[2], vocab_all_values[0]])
         factory_wrp = getUtility(IVocabularyFactory, "collective.contact.plonegroup.selected_organization_services")
         vocab_wrp = factory_wrp(self.portal)
         self.assertEqual(len(vocab_wrp), 3)
         # values are sorted by title
-        self.assertListEqual(sorted([v.value for v in vocab_wrp]), sorted(registry[ORGANIZATIONS_REGISTRY]))
+        self.assertListEqual(sorted([v.value for v in vocab_wrp]), sorted(get_registry_organizations()))
         self.assertListEqual([v.title for v in vocab_wrp], ['Department 1', 'Department 2'])
         self.assertEqual(vocab_wrp.getTerm(vocab_all_values[1]).title, 'Department 1 - Service 1')