+++ src/collective/contact/plonegroup/tests/test_subscribers.py
@@ -1,20 +1,20 @@
 # -*- coding: utf-8 -*-
 """Setup/installation tests for this package."""
 
-from collective.contact.plonegroup.config import FUNCTIONS_REGISTRY
-from collective.contact.plonegroup.config import ORGANIZATIONS_REGISTRY
+from collective.contact.plonegroup.config import DEFAULT_DIRECTORY_ID
 from collective.contact.plonegroup.config import PLONEGROUP_ORG
+from collective.contact.plonegroup.config import set_registry_functions
+from collective.contact.plonegroup.config import set_registry_organizations
 from collective.contact.plonegroup.interfaces import INotPloneGroupContact
 from collective.contact.plonegroup.interfaces import IPloneGroupContact
 from collective.contact.plonegroup.subscribers import group_deleted
 from collective.contact.plonegroup.testing import IntegrationTestCase
+from collective.contact.plonegroup.utils import get_own_organization
 from plone import api
 from plone.app.linkintegrity.exceptions import LinkIntegrityNotificationException
 from plone.app.linkintegrity.interfaces import ILinkIntegrityInfo
-from plone.registry.interfaces import IRegistry
 from Products.statusmessages.interfaces import IStatusMessage
 from zExceptions import Redirect
-from zope.component import getUtility
 
 
 class TestSubscribers(IntegrationTestCase):
@@ -24,17 +24,15 @@ class TestSubscribers(IntegrationTestCase):
         """Custom shared utility setup for tests."""
         self.portal = self.layer['portal']
         # Organizations creation
-        self.portal.invokeFactory('directory', 'contacts')
-        self.portal['contacts'].invokeFactory('organization', PLONEGROUP_ORG, title='My organization')
-        own_orga = self.portal['contacts'][PLONEGROUP_ORG]
+        self.portal.invokeFactory('directory', DEFAULT_DIRECTORY_ID)
+        self.portal[DEFAULT_DIRECTORY_ID].invokeFactory('organization', PLONEGROUP_ORG, title='My organization')
+        own_orga = get_own_organization()
         own_orga.invokeFactory('organization', 'department1', title='Department 1')
         own_orga.invokeFactory('organization', 'department2', title='Department 2')
         self.contacts = [own_orga['department1'], own_orga['department2']]
 
-        self.registry = getUtility(IRegistry)
-        self.registry[ORGANIZATIONS_REGISTRY] = [c.UID() for c in self.contacts]
-        self.registry[FUNCTIONS_REGISTRY] = [
-            {'fct_title': u'Director', 'fct_id': u'director', 'fct_orgs': []}]
+        set_registry_organizations([c.UID() for c in self.contacts])
+        set_registry_functions([{'fct_title': u'Director', 'fct_id': u'director', 'fct_orgs': []}])
 
         self.portal.invokeFactory('acontent',
                                   'acontent1',
@@ -47,7 +45,8 @@ class TestSubscribers(IntegrationTestCase):
 
     def test_plonegroupOrganizationRemoved_1(self):
         """ We cannot remove an organization selected in settings and used in an object """
-        view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
+        view = self.portal.restrictedTraverse(
+            '{0}/{1}/department1/delete_confirmation'.format(DEFAULT_DIRECTORY_ID, PLONEGROUP_ORG))
         self.assertRaises(LinkIntegrityNotificationException, view.render)
         storage = ILinkIntegrityInfo(view.REQUEST)
         breaches = storage.getIntegrityBreaches()
@@ -56,8 +55,9 @@ class TestSubscribers(IntegrationTestCase):
 
     def test_plonegroupOrganizationRemoved_2(self):
         """ We cannot remove an organization no more selected in settings and used in an object """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
-        view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+        set_registry_organizations([self.contacts[0].UID()])  # unselects the contact
+        view = self.portal.restrictedTraverse(
+            '{0}/{1}/department2/delete_confirmation'.format(DEFAULT_DIRECTORY_ID, PLONEGROUP_ORG))
         self.assertRaises(LinkIntegrityNotificationException, view.render)
         storage = ILinkIntegrityInfo(view.REQUEST)
         breaches = storage.getIntegrityBreaches()
@@ -66,15 +66,17 @@ class TestSubscribers(IntegrationTestCase):
 
     def test_plonegroupOrganizationRemoved_3(self):
         """ We can remove an organization no more selected in settings and no more used in an object """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
+        set_registry_organizations([self.contacts[0].UID()])  # unselects the contact
         self.portal['acontent2'].pg_organization = None
-        self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+        self.portal.restrictedTraverse(
+            '{0}/{1}/department2/delete_confirmation'.format(DEFAULT_DIRECTORY_ID, PLONEGROUP_ORG))
 
     def test_plonegroupOrganizationRemoved_4(self):
         """ We cannot remove an organization selected in settings and used in an object as dict or list """
         # set uid in dict
         self.portal['acontent1'].pg_organization = {'uid': self.contacts[0].UID()}
-        view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
+        view = self.portal.restrictedTraverse(
+            '{0}/{1}/department1/delete_confirmation'.format(DEFAULT_DIRECTORY_ID, PLONEGROUP_ORG))
         self.assertRaises(LinkIntegrityNotificationException, view.render)
         storage = ILinkIntegrityInfo(view.REQUEST)
         breaches = storage.getIntegrityBreaches()
@@ -82,7 +84,8 @@ class TestSubscribers(IntegrationTestCase):
         self.assertSetEqual(breaches[self.contacts[0]], set([self.portal['acontent1']]))
         # set uid in list
         self.portal['acontent2'].pg_organization = [self.contacts[1].UID()]
-        view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+        view = self.portal.restrictedTraverse(
+            '{0}/{1}/department2/delete_confirmation'.format(DEFAULT_DIRECTORY_ID, PLONEGROUP_ORG))
         self.assertRaises(LinkIntegrityNotificationException, view.render)
         storage = ILinkIntegrityInfo(view.REQUEST)
         breaches = storage.getIntegrityBreaches()
@@ -96,20 +99,20 @@ class TestSubscribers(IntegrationTestCase):
 
     def test_plonegroup_contact_transition_2(self):
         """ We cannot deactivate an organization used in objects """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[1].UID()]  # unselects the contact
+        set_registry_organizations([self.contacts[1].UID()])  # unselects the contact
         self.assertRaises(Redirect, api.content.transition, obj=self.contacts[0], transition='deactivate')
         self.assertEqual(api.content.get_state(obj=self.contacts[0]), 'active')
 
     def test_plonegroup_contact_transition_3(self):
         """ We can deactivate an organization not at all used """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[1].UID()]  # unselects the contact
+        set_registry_organizations([self.contacts[1].UID()])  # unselects the contact
         self.portal['acontent1'].pg_organization = None
         api.content.transition(obj=self.contacts[0], transition='deactivate')
         self.assertEqual(api.content.get_state(obj=self.contacts[0]), 'deactivated')
 
     def test_mark_organization(self):
         """ We test marker interfaces """
-        contacts = self.portal.contacts
+        contacts = self.portal.get(DEFAULT_DIRECTORY_ID)
         pg_org = contacts[PLONEGROUP_ORG]
         self.assertTrue(IPloneGroupContact.providedBy(self.contacts[0]))
         self.assertFalse(INotPloneGroupContact.providedBy(self.contacts[0]))